requirements.txt
streamlit
fastapi
uvicorn[standard]
pydantic
pyyaml
matplotlib
pandas
watchdog
